requirements.txt
fastapi==0.112.0
fastapi-cli==0.0.5
pymongo==4.8.0
python-dotenv==1.0.1
pydantic~=2.8.2